Promo code deep linking instructions
Added a link to an article explaining how to create a deep link with the app promo code for 1-click downloading.  Full disclosure - this links to a blog I help maintain however I feel that this information is valuable to app developers and is in the spirit of this document.